Egg.Farm

Documentation

data FarmStatus :: Nat -> ([Nat], Nat) -> Nat -> Nat -> Type where

Constructors

FarmStatus :: {..} -> FarmStatus eggs '(tiers, epic) habs vehicles

Fields

_fsEgg :: Finite eggs

_fsResearch :: ResearchStatus tiers epic

_fsHabs :: HabStatus habs

_fsDepot :: SomeDepotStatus vehicles

_fsHatchery :: Double

_fsBocks :: Bock

_fsGoldenEggs :: GoldenEgg

_fsSoulEggs :: SoulEgg

_fsVideoBonus :: Maybe Double

in seconds

_fsPrestEarnings :: Bock

fsEgg :: Lens (FarmStatus e1 '(t, g) h v) (FarmStatus e2 '(t, g) h v) (Finite e1) (Finite e2)

fsResearch :: Lens (FarmStatus e '(t1, g1) h v) (FarmStatus e '(t2, g2) h v) (ResearchStatus t1 g1) (ResearchStatus t2 g2)

fsHabs :: Lens (FarmStatus e '(t, g) h1 v) (FarmStatus e '(t, g) h2 v) (HabStatus h1) (HabStatus h2)

fsHatchery :: Lens' (FarmStatus e '(t, g) h v) Double

fsDepot :: Lens (FarmStatus e '(t, g) h v1) (FarmStatus e '(t, g) h v2) (SomeDepotStatus v1) (SomeDepotStatus v2)

fsBocks :: Lens' (FarmStatus e '(t, g) h v) Bock

fsGoldenEggs :: Lens' (FarmStatus e '(t, g) h v) GoldenEgg

fsSoulEggs :: Lens' (FarmStatus e '(t, g) h v) SoulEgg

fsVideoBonus :: Lens' (FarmStatus e '(t, g) h v) (Maybe Double)

Enforces only postive Just

fsPrestEarnings :: Lens' (FarmStatus e '(t, g) h v) Bock

data IncomeChange

Constructors

ICHabFill (Fin N4)

ICDepotFill

ICVideoDoublerExpire

_ICHabFill :: Prism' IncomeChange (Fin N4)

_ICDepotFill :: Prism' IncomeChange ()

_ICVideoDoublerExpire :: Prism' IncomeChange ()

data UpgradeError

Constructors

UELocked

UERegression

_UELocked :: Prism' UpgradeError ()

_UERegression :: Prism' UpgradeError ()

data HatchError

Constructors

HEHatcheryLow

Fields

_heHatchery :: Double

HEHabsFull

Fields

_heHabitats :: Double

heHabitats :: Traversal' HatchError Double

heHatchery :: Traversal' HatchError Double

data Drone

Constructors

NormalDrone (Fin N3)

EliteDrone

_NormalDrone :: Prism' Drone (Fin N3)

_EliteDrone :: Prism' Drone ()

initFarmStatus :: (KnownNat eggs, KnownNat habs, KnownNat vehicles, 1 <= eggs, 1 <= habs, 1 <= vehicles) => GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles

farmEggValue :: KnownNat eggs => GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> Bock

Egg value

farmLayingRate :: KnownNat vehicles => GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> Double

Egg rate in eggs per second.

Multiplies farmLayingRatePerChicken by the number of chickens, and caps it based on the depot capacity.

Also gives remaining capacity of depots, with Nothing if depots are at full capacity.

farmLayingRatePerChicken :: GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> Double

How many eggs are laid per second, per chicken.

farmLayingRateAvailability :: KnownNat vehicles => GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> (Double, Maybe Double)

Egg rate in eggs per second.

Multiplies farmLayingRatePerChicken by the number of chickens, and caps it based on the depot capacity.

Also gives remaining capacity of depots, with Nothing if depots are at full capacity.

hatcheryRefillRate :: GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> Double

Chickens per second

hatcheryCapacity :: GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> Double

refillHatcheries

Arguments

:: GameData eggs '(tiers, epic) habs vehicles

-> Double
dt

-> FarmStatus eggs '(tiers, epic) habs vehicles

-> FarmStatus eggs '(tiers, epic) habs vehicles

farmIncomeNoBonus :: (KnownNat eggs, KnownNat vehicles) => GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> Bock

Total income (bocks per second), no bonuses

farmIncome :: (KnownNat eggs, KnownNat vehicles) => GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> Bock

Total income (bocks per second), with bonuses

farmIncomeDT

Arguments

:: (KnownNat eggs, KnownNat habs, KnownNat vehicles)

=> GameData eggs '(tiers, epic) habs vehicles

-> IsCalm

-> FarmStatus eggs '(tiers, epic) habs vehicles

-> Double
small time step (dt^3 = 0)

-> Bock

Total bock income over a given small-ish period of time, assuming no habs get filled up or depots get filled up in this time interval.

Time step should be small (dt^3 = 0) enough to avoid discontinuities, but this function still takes into account quadratic growth due to dt^2 terms.

habSize(t) = initHabSize + growthRate * t incomerate(t) = capped (habSize(t) * layingRate) * eggValue

Assuming capped:

incomerate(t) = (initHabSize + growthRate * t) * layingRate * eggValue * bonuses totalincome = [initHabSize * dt + 1/2 growthRate * dt^2] * layingRate * eggValue * bonuses

Assuming uncapped:

incomerate(t) = capped (initHabSize * layingRate) * eggValue totalincome = dt * capped (initHabSize * layingRate) * eggValue

farmIncomeDTInv

Arguments

:: (KnownNat eggs, KnownNat habs, KnownNat vehicles)

=> GameData eggs '(tiers, epic) habs vehicles

-> IsCalm

-> FarmStatus eggs '(tiers, epic) habs vehicles

-> Bock
bock goal

-> Maybe Double

Find the time to the given goal, assuming constant growth rate and no discontinuities.

Assuming uncapped:

totalincome = [initHabSize * dt + 1/2 growthRate * dt^2] * layingRate * eggValue * bonuses

so we hav:

12 growthRate dt^2 + initHabSize dt - totalincome (layingRate * eggValue * bonuses) = 0

Assuming capped:

totalincome = capped (initHabSize dt * layingRate) * evvGalue * bonuses

so:

dt = totalincome / capped rate

Assumes egg value is not zero.

farmDepotCapacity :: KnownNat vehicles => GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> Double

Total depot capacity of farm, in eggs per second

farmValue :: (KnownNat eggs, KnownNat habs, KnownNat vehicles) => GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> Bock

Farm value

Formula is current hardcoded.

Credit upatashu:

https://www.reddit.com/r/EggsInc/comments/63pfye/patashus_egg_inc_guide_xpost_from_regginc/

TODO: figure out if income should be capped by depot or not

bonusingSoulEggs :: GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> Iso' Bock Bock

Soul egg bonus

videoDoubling :: GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> Iso' Bock Bock

Apply video doubling bonus

farmBonuses :: GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> Bonuses

Bonuses from a given GameData and FarmStatus

stepFarm

Arguments

:: (KnownNat eggs, KnownNat habs, KnownNat vehicles)

=> GameData eggs '(tiers, epic) habs vehicles

-> IsCalm

-> Double
time interval to step

-> FarmStatus eggs '(tiers, epic) habs vehicles

-> FarmStatus eggs '(tiers, epic) habs vehicles

Step the farm over a given time interval

Should be += integral (R c(t)) dt where c'(t) = rt

stepFarmDT

Arguments

:: (KnownNat eggs, KnownNat habs, KnownNat vehicles)

=> GameData eggs '(tiers, epic) habs vehicles

-> IsCalm

-> Double
small (infinitessimal) time interval to step

-> FarmStatus eggs '(tiers, epic) habs vehicles

-> FarmStatus eggs '(tiers, epic) habs vehicles

Step farm over a SMALL (infinitessimal) time step. Assumes that nothing discontinuous changes during the time inveral, including things like:

habs filling up

depots filling up

Useful for stepping continuous simulations.

waitUntilBocks

Arguments

:: (KnownNat eggs, KnownNat habs, KnownNat vehicles)

=> GameData eggs '(tiers, epic) habs vehicles

-> IsCalm

-> Bock
goal

-> FarmStatus eggs '(tiers, epic) habs vehicles

-> WaitTilRes I (FarmStatus eggs '(tiers, epic) habs vehicles)

Wait until bocks.

waitTilFarmPop

Arguments

:: (KnownNat eggs, KnownNat habs, KnownNat vehicles)

=> GameData eggs '(tiers, epic) habs vehicles

-> IsCalm

-> Natural
goal

-> FarmStatus eggs '(tiers, epic) habs vehicles

-> WaitTilRes (Either (FarmStatus eggs '(tiers, epic) habs vehicles)) (FarmStatus eggs '(tiers, epic) habs vehicles)

Wait until population reaches a given point.

If Left returned, means that hab maxed out.

waitTilDepotFull :: (KnownNat eggs, KnownNat habs, KnownNat vehicles) => GameData eggs '(tiers, epic) habs vehicles -> IsCalm -> FarmStatus eggs '(tiers, epic) habs vehicles -> WaitTilRes (Either (FarmStatus eggs '(tiers, epic) habs vehicles)) (FarmStatus eggs '(tiers, epic) habs vehicles)

Time until depots are full

If Left returned, means that hab maxed out.

hatchChickens

Arguments

:: KnownNat habs

=> GameData eggs '(tiers, epic) habs vehicles

-> Natural
number of chickens

-> FarmStatus eggs '(tiers, epic) habs vehicles

-> Either HatchError (FarmStatus eggs '(tiers, epic) habs vehicles)

upgradeEgg :: (KnownNat eggs, KnownNat habs, KnownNat vehicles, 1 <= habs, 1 <= vehicles) => GameData eggs '(tiers, epic) habs vehicles -> Finite eggs -> FarmStatus eggs '(tiers, epic) habs vehicles -> Either UpgradeError (FarmStatus eggs '(tiers, epic) habs vehicles)

Upgrade your egg!

eggUpgrades :: (KnownNat eggs, KnownNat habs, KnownNat vehicles) => GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> (Vector eggs :.: Either UpgradeError) (Finite eggs)

Vector of possible upgrades

eggsVisible :: (KnownNat eggs, KnownNat habs, KnownNat vehicles) => GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> (Vector eggs :.: Maybe) (Finite eggs)

Vector of visible eggs

prestigeFarm :: (KnownNat eggs, KnownNat habs, KnownNat vehicles, 1 <= habs, 1 <= vehicles) => GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles

Prestige! Resets farm, increments soul eggs.

TODO: only available if pBonus > 0

watchVideo :: GameData eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles -> Either Double (FarmStatus eggs '(tiers, epic) habs vehicles)

Watch Video Doubler video.

Returns Left if video doubler refresh limit has not yet been reached, with time left.

popDrone :: (KnownNat eggs, KnownNat habs, KnownNat vehicles) => GameData eggs '(tiers, epic) habs vehicles -> Drone -> FarmStatus eggs '(tiers, epic) habs vehicles -> FarmStatus eggs '(tiers, epic) habs vehicles