Egg.Habitat

Documentation

data Hab

Constructors

Hab

Fields

_habName :: Text

_habBaseCapacity :: Natural

_habCosts :: Vec N4 Bock

habName :: Lens' Hab Text

habBaseCapacity :: Lens' Hab Natural

habCosts :: Lens' Hab (Vec N4 Bock)

newtype HabData habs

Constructors

HabData

Fields

_hdHabs :: Vector habs Hab

_HabData :: forall habs habs. Iso (HabData habs) (HabData habs) (Vector habs Hab) (Vector habs Hab)

type SomeHabData = DSum Sing HabData

data HabStatus habs

Constructors

HabStatus

Fields

_hsSlots :: Vec N4 (Maybe (Finite habs))

_hsPop :: Vec N4 Bock

_HabStatus :: Iso' (HabStatus habs) (Vec N4 (Maybe (Finite habs), Double))

hsSlots :: forall habs habs. Lens (HabStatus habs) (HabStatus habs) (Vec N4 (Maybe (Finite habs))) (Vec N4 (Maybe (Finite habs)))

hsPop :: forall habs. Lens' (HabStatus habs) (Vec N4 Bock)

data IsCalm

Constructors

NotCalm

Calm

_NotCalm :: Prism' IsCalm ()

_Calm :: Prism' IsCalm ()

initHabStatus :: (KnownNat habs, 1 <= habs) => HabStatus habs

Initial HabStatus to start off the game.

habAt :: KnownNat habs => HabData habs -> HabStatus habs -> Fin N4 -> Maybe Hab

Hab at the given slot

totalChickens :: HabData habs -> HabStatus habs -> Double

Total number of chickens

baseHabCapacity :: forall habs. KnownNat habs => HabData habs -> HabStatus habs -> Natural

Total base capacity of all slots.

baseHabCapacities :: forall habs. KnownNat habs => HabData habs -> HabStatus habs -> Vec N4 Natural

Base capacity of each slot.

totalHabCapacity :: forall habs. KnownNat habs => HabData habs -> Bonuses -> HabStatus habs -> Natural

Total capacity of all hatcheries, factoring in bonuses.

habCapacity :: Bonuses -> Hab -> Natural

Get the actual capacity for a given hab, with given bonuses.

fullHabs :: forall habs. KnownNat habs => HabData habs -> Bonuses -> HabStatus habs -> Vec N4 Bool

Which habs are full?

availableSpace :: forall habs. KnownNat habs => HabData habs -> Bonuses -> HabStatus habs -> Vec N4 (Maybe Double, Natural)

Gives total avaiable space of each hab (Nothing if hab full), and also the capacity of each hab.

habCapacities :: forall habs. KnownNat habs => HabData habs -> Bonuses -> HabStatus habs -> Vec N4 Natural

Capacities of all slots, factoring in bonuses.

habHistory :: HabStatus habs -> Map (Finite habs) (Fin N5)

How many of each hab is currently owned. If key is not found, zero purchases is implied.

habPrice :: KnownNat habs => HabData habs -> HabStatus habs -> Finite habs -> Maybe Bock

Get the BASE price of a given hab, if a purchase were to be made. Does not check if purchase is legal (see upgradeHab).

upgradeHab :: KnownNat habs => HabData habs -> Bonuses -> Fin N4 -> Finite habs -> HabStatus habs -> Maybe (Bock, HabStatus habs)

Purchase a hab upgrade. Returns cost and new hab status, if purchase is valid.

Purchase is invalid if purchasing a hab in a slot where a greater hab is already purchased.

internalHatcheryRate :: Bonuses -> IsCalm -> Double

Compute the base internal hatchery rate (chickens per second per hatchery) from bonuses.

data WaitError

Constructors

WENoInternalHatcheries

WEMaxHabCapacity

_WENoInternalHatcheries :: Prism' WaitError ()

_WEMaxHabCapacity :: Prism' WaitError ()

waitTilNextFilled :: forall habs. KnownNat habs => HabData habs -> Bonuses -> IsCalm -> HabStatus habs -> Either WaitError ((Fin N4, Double), (HabStatus habs, Double))

Calculate the time until the next hab is full, and return the updated habs after that time. Returns Nothing if all habs are full.

Also returns the current total fill rate per hatchery, taking into account all bonuses.

Conclusions:

No internal hatcheries

All habs are already full.

Hab n filled after a given amount of time.

stepHabs

Arguments

:: KnownNat habs

=> HabData habs

-> Bonuses

-> IsCalm

-> Double
time interval to step

-> HabStatus habs
initial status

-> HabStatus habs

Steps the HabStatus over the given time interval.

data WaitTilRes habs

Constructors

WaitTilSuccess Double (HabStatus habs)

MaxHabPopIn Double

NoWait

NoInternalHatcheries

_WaitTilSuccess :: forall habs habs. Prism (WaitTilRes habs) (WaitTilRes habs) (Double, HabStatus habs) (Double, HabStatus habs)

_MaxHabPopIn :: forall habs. Prism' (WaitTilRes habs) Double

_NoWait :: forall habs. Prism' (WaitTilRes habs) ()

_NoInternalHatcheries :: forall habs. Prism' (WaitTilRes habs) ()

wtrRes :: forall habs. Traversal' (WaitTilRes habs) Double

wtrStatus :: forall habs habs. Traversal (WaitTilRes habs) (WaitTilRes habs) (HabStatus habs) (HabStatus habs)

waitTilPop :: forall habs. KnownNat habs => HabData habs -> Bonuses -> IsCalm -> Natural -> HabStatus habs -> WaitTilRes habs

Time until a given population is reached.

fillTimes :: forall habs. KnownNat habs => HabData habs -> Bonuses -> IsCalm -> HabStatus habs -> Either WaitError (Vec N4 Double)

Times for each hab to be filled.