Egg.Research

Documentation

data BonusAmount

Constructors

BAIncrement Double
Bonus adds an absolute amount. Scales by multiplication, and compounds by addition.

BAPercent Double
Bonus adds a percentage point multiplier. Scales by multiplication (three 10% buffs is a 15% buff) and compounds by appropriate multiplication.

BAMultiplier Double
Bonus multiplies by a given ratio. Scales by exponentiation (three 2x buffs is a 8x buff) and compounds by multiplication.

class AsBonusAmount r where

Methods

_BonusAmount :: Prism' r BonusAmount

_BAIncrement :: Prism' r Double

_BAPercent :: Prism' r Double

_BAMultiplier :: Prism' r Double

data BonusType

Constructors

BTBuildCosts

BTDroneRewards

BTEggValue

BTFarmValue

BTFleetSize

BTHabCapacity

BTHatcheryCapacity

BTHatcheryRate

BTHoldingHatch

BTHoverVehicleCapacity

BTInternalHatchery

BTInternalHatcheryCalm

BTInternalHatcherySharing

BTLayingRate

BTLongWarpTime

BTMaxRunningBonus

BTPrestigeEggs

BTResearchCosts

BTRunningBonus

BTSiloQuality

BTSoulEggBonus

BTVehicleCapacity

BTVehicleCosts

BTVideoDoublerTime

BTVehicleSpeed

class AsBonusType r where

Methods

_BonusType :: Prism' r BonusType

_BTBuildCosts :: Prism' r ()

_BTDroneRewards :: Prism' r ()

_BTEggValue :: Prism' r ()

_BTFarmValue :: Prism' r ()

_BTFleetSize :: Prism' r ()

_BTHabCapacity :: Prism' r ()

_BTHatcheryCapacity :: Prism' r ()

_BTHatcheryRate :: Prism' r ()

_BTHoldingHatch :: Prism' r ()

_BTHoverVehicleCapacity :: Prism' r ()

_BTInternalHatchery :: Prism' r ()

_BTInternalHatcheryCalm :: Prism' r ()

_BTInternalHatcherySharing :: Prism' r ()

_BTLayingRate :: Prism' r ()

_BTLongWarpTime :: Prism' r ()

_BTMaxRunningBonus :: Prism' r ()

_BTPrestigeEggs :: Prism' r ()

_BTResearchCosts :: Prism' r ()

_BTRunningBonus :: Prism' r ()

_BTSiloQuality :: Prism' r ()

_BTSoulEggBonus :: Prism' r ()

_BTVehicleCapacity :: Prism' r ()

_BTVehicleCosts :: Prism' r ()

_BTVideoDoublerTime :: Prism' r ()

_BTVehicleSpeed :: Prism' r ()

newtype Bonuses

Constructors

Bonuses

Fields

_bMap :: Map BonusType [BonusAmount]

_Bonuses :: Iso' Bonuses (Map BonusType [BonusAmount])

data Research a

Constructors

Research

Fields

_rName :: Text

_rDescription :: Text

_rBaseBonuses :: Bonuses

_rCosts :: Either Natural (Vector a)

Left if no cost data, just giving the maximum level. Right with cost data.

class HasResearch c a | c -> a where

Minimal complete definition

research

Methods

research :: Lens' c (Research a)

rBaseBonuses :: Lens' c Bonuses

rCosts :: Lens' c (Either Natural (Vector a))

rDescription :: Lens' c Text

rName :: Lens' c Text

data ResearchTier :: Nat -> Type where

Constructors

ResearchTier :: {..} -> ResearchTier n

Fields

_rtUnlock :: Natural

_rtTechs :: Vector n (Research Bock)

rtUnlock :: forall a. Lens' (ResearchTier a) Natural

rtTechs :: forall a a. Lens (ResearchTier a) (ResearchTier a) (Vector a (Research Bock)) (Vector a (Research Bock))

type SomeResearchTier = DSum Sing ResearchTier

data ResearchData :: [Nat] -> Nat -> Type where

Constructors

ResearchData :: {..} -> ResearchData tiers epic

Fields

_rdCommon :: Prod ResearchTier tiers

_rdEpic :: Vector epic (Research GoldenEgg)

rdCommon :: forall a b a. Lens (ResearchData a b) (ResearchData a b) (Prod ResearchTier a) (Prod ResearchTier a)

rdEpic :: forall a b b. Lens (ResearchData a b) (ResearchData a b) (Vector b (Research GoldenEgg)) (Vector b (Research GoldenEgg))

type SomeResearchData = DSum Sing (Uncur ResearchData)

data ResearchStatus :: [Nat] -> Nat -> Type where

Constructors

ResearchStatus :: {..} -> ResearchStatus tiers epic

Fields

_rsCommon :: Prod (Flip Vector Natural) tiers

_rsEpic :: Vector epic Natural

rsCommon :: forall a b a. Lens (ResearchStatus a b) (ResearchStatus a b) (Prod (Flip Vector Natural :: Nat -> Type) a) (Prod (Flip Vector Natural :: Nat -> Type) a)

rsEpic :: forall a b b. Lens (ResearchStatus a b) (ResearchStatus a b) (Vector b Natural) (Vector b Natural)

type SomeResearchStatus = DSum Sing (Uncur ResearchStatus)

data ResearchIx :: [Nat] -> Nat -> Type -> Type where

A safe index for a given research item, usable with ResearchData and ResearchStatus.

Constructors

RICommon :: Sum Finite tiers -> ResearchIx tiers epic Bock

RIEpic :: Finite epic -> ResearchIx tiers epic GoldenEgg

_RICommon :: Iso (ResearchIx t1 epic Bock) (ResearchIx t2 epic Bock) (Sum Finite t1) (Sum Finite t2)

_RIEpic :: Iso (ResearchIx tiers e1 GoldenEgg) (ResearchIx tiers e2 GoldenEgg) (Finite e1) (Finite e2)

data ResearchError

Constructors

REMaxedOut

RELocked

_REMaxedOut :: Prism' ResearchError ()

_RELocked :: Prism' ResearchError ()

scaleAmount :: Natural -> BonusAmount -> BonusAmount

Scales a bonus amount by a "level".

hasEffect :: BonusAmount -> Bool

Tells whether or not a bonus amount creates any actual effect.

bonusEffect :: [BonusAmount] -> Double -> Double

Apply a list of bonuses (from left to right) to a value.

bonusEffectFor :: Bonuses -> BonusType -> Double -> Double

Apply bonuses for a given type on a value.

bonusing :: [BonusAmount] -> Iso' Double Double

Iso on a value under a list of bonuses (from left to right).

bonusingFor :: Bonuses -> BonusType -> Iso' Double Double

Iso on a value under the bonuses of a given bonus type.

emptyBonuses :: Bonuses

No bonuses

bonuses :: Map BonusType [BonusAmount] -> Bonuses

A "smart constructor" that clears bonuses with no effect.

maxLevel :: Research a -> Natural

Maximum level for a given research.

emptyResearchStatus :: ResearchData tiers epic -> ResearchStatus tiers epic

Empty status (no research).

researchBonuses :: Research a -> Natural -> Bonuses

Bonuses from a given Research at a given level. Assumes no "maximum level".

totalBonuses :: ResearchData tiers epic -> ResearchStatus tiers epic -> Bonuses

Total bonuses from a given ResearchStatus.

researchCount :: ResearchStatus tiers epic -> Natural

How many common techs have been researched?

Used for opening tiers.

legalTiers :: forall tiers epic. ResearchData tiers epic -> ResearchStatus tiers epic -> Prod (C Bool) tiers

foldResearch :: Monoid b => (Either (Research Bock) (Research GoldenEgg) -> Natural -> b) -> ResearchData tiers epic -> ResearchStatus tiers epic -> b

Zips together all research data and research statuses into an accumulator.

purchaseResearch :: forall tiers epic a. (KnownNat epic, SingI tiers) => ResearchData tiers epic -> ResearchIx tiers epic a -> ResearchStatus tiers epic -> Either ResearchError (a, ResearchStatus tiers epic)

Purchase research at a given index, incrementing the counter in the ResearchStatus. Returns Nothing if research is already maxed out.

researchIxNum :: ResearchIx tiers epic a -> Wit (Num a)

Get a Num instance from a ResearchIx.

researchIxData :: (KnownNat epic, SingI tiers) => ResearchIx tiers epic a -> Lens' (ResearchData tiers epic) (Research a)

A lens into a ResearchData, given the appropriate index.

researchIxStatus :: (KnownNat epic, SingI tiers) => ResearchIx tiers epic a -> Lens' (ResearchStatus tiers epic) Natural

A lens into a ResearchStatus, given the appropriate index.

researchIxStatusLegal :: (SingI tiers, KnownNat epic) => ResearchData tiers epic -> ResearchIx tiers epic a -> Traversal' (ResearchStatus tiers epic) Natural

Traversal into a given index of a ResearchStatus if the item is in a legal tier.

Only a legal traversal if the mapping function doesn't change the legal status.

withSomeResearch :: DSum Sing (Uncur res) -> (forall tiers epic. (KnownNat epic, SingI tiers) => res tiers epic -> r) -> r

researchIxesCommon :: SingI tiers => Prod (Flip Vector (ResearchIx tiers epic Bock)) tiers

All ResearchIx into common researches.

researchIxesEpic :: KnownNat epic => Vector epic (ResearchIx tiers epic GoldenEgg)

All ResearchIx into epic researches.

legalResearchIxesCommon :: forall tiers epic. SingI tiers => ResearchData tiers epic -> ResearchStatus tiers epic -> Prod (Flip Vector (Either ResearchError (ResearchIx tiers epic Bock))) tiers

All legal ResearchIx for common research.

legalResearchIxesEpic :: forall tiers epic. KnownNat epic => ResearchData tiers epic -> ResearchStatus tiers epic -> (Vector epic :.: Maybe) (ResearchIx tiers epic GoldenEgg)

All legal ResearchIx for epic research.

Nothing implies research is maxed-out.

Why?

rdrsCommon :: Lens' ((Uncur ResearchData :&: Uncur ResearchStatus) (tiers # epic)) (Prod (ResearchTier :&: Flip Vector Natural) tiers)

Lens into both the research data and research status common slots together.

rdrsEpic :: KnownNat epic => Lens' ((Uncur ResearchData :&: Uncur ResearchStatus) (tiers # epic)) (Vector epic (Research GoldenEgg, Natural))

Lens into both the research data and research status epic slots together.

researchIx :: (SingI tiers, KnownNat epic) => ResearchIx tiers epic a -> Lens' ((Uncur ResearchData :&: Uncur ResearchStatus) (tiers # epic)) (Research a, Natural)

Lens into both a ResearchData and ResearchStatus together, from a given index.