Egg.Vehicle

Documentation

data Vehicle

Constructors

Vehicle

Fields

_vName :: Text

_vBaseCapacity :: Natural

eggs per minute

_vCosts :: Vector Double

vName :: Lens' Vehicle Text

vBaseCapacity :: Lens' Vehicle Natural

vCosts :: Lens' Vehicle (Vector Double)

newtype VehicleData vs

Constructors

VehicleData

Fields

_vdVehicles :: Vector vs Vehicle

_VehicleData :: forall vs vs. Iso (VehicleData vs) (VehicleData vs) (Vector vs Vehicle) (Vector vs Vehicle)

type SomeVehicleData = DSum Sing VehicleData

data DepotStatus vs slots

Constructors

DepotStatus

Fields

_dsSlots :: Vec slots (Set (Finite vs))

_DepotStatus :: forall vs slots vs slots. Iso (DepotStatus vs slots) (DepotStatus vs slots) (Vec slots (Set (Finite vs))) (Vec slots (Set (Finite vs)))

initDepotStatus :: KnownNat vs => DepotStatus vs N4

Initial DepotStatus to start off the game.

upgradeDepo :: forall vs n. Known Nat n => DepotStatus vs n -> DepotStatus vs (S n)

Add another empty depot slot

baseCapacity :: forall vs slots. KnownNat vs => VehicleData vs -> DepotStatus vs slots -> Double

Total base capacity of all slots, in eggs per second.

totalCapacity :: forall vs slots. KnownNat vs => VehicleData vs -> Bonuses -> DepotStatus vs slots -> Double

Total capacity of all vehicles, factoring in bonuses.

vehicleHistory :: Known Nat slots => DepotStatus vs slots -> Map (Finite vs) (Fin (S slots))

How many of each vehicle has been purchased so far. If key is not found, zero purchases is implied.

vehiclePrice :: forall vs slots. (KnownNat vs, Known Nat slots) => VehicleData vs -> DepotStatus vs slots -> Finite vs -> Maybe Double

Get the BASE price of a given vehicle, if a purchase were to be made. Does not check if purchase is legal (see upgradeVehicle).

upgradeVehicle :: (KnownNat vs, Known Nat slots) => VehicleData vs -> Bonuses -> Fin slots -> Finite vs -> DepotStatus vs slots -> Maybe (Double, DepotStatus vs slots)

Purchase a vehicle upgrade. Returns cost and new depot status, if purchase is valid.

Purchase is invalid if purchasing a vehicle in a slot where a greater vehicle is already purchased.