Egg.Vehicle

Documentation

data Vehicle

Constructors

Vehicle

Fields

_vName :: Text

_vBaseCapacity :: Natural

eggs per minute

_vCosts :: Vector Bock

vName :: Lens' Vehicle Text

vBaseCapacity :: Lens' Vehicle Natural

vCosts :: Lens' Vehicle (Vector Bock)

newtype VehicleData vs

Constructors

VehicleData

Fields

_vdVehicles :: Vector vs Vehicle

_VehicleData :: forall vs vs. Iso (VehicleData vs) (VehicleData vs) (Vector vs Vehicle) (Vector vs Vehicle)

type SomeVehicleData = DSum Sing VehicleData

data DepotStatus vs slots

Constructors

DepotStatus

Fields

_dsSlots :: Map (Finite slots) (Finite vs)

_DepotStatus :: forall vs slots vs slots. Iso (DepotStatus vs slots) (DepotStatus vs slots) (Map (Finite slots) (Finite vs)) (Map (Finite slots) (Finite vs))

data SomeDepotStatus vs

_SomeDepotStatus :: Functor f => Bonuses -> (forall slots. KnownNat slots => LensLike' f (DepotStatus vs slots) b) -> LensLike' f (SomeDepotStatus vs) b

data SomeVehicleUpgradeError vs

Constructors

SVUERegression

Fields

_svuePrevious :: Finite vs

SVUENoSlot

_SVUERegression :: forall vs vs. Prism (SomeVehicleUpgradeError vs) (SomeVehicleUpgradeError vs) (Finite vs) (Finite vs)

_SVUENoSlot :: forall vs. Prism' (SomeVehicleUpgradeError vs) ()

withSomeDepotStatus :: Bonuses -> SomeDepotStatus vs -> (forall slots. KnownNat slots => DepotStatus vs slots -> r) -> r

initDepotStatus :: (KnownNat vs, KnownNat slots, 1 <= vs, 1 <= slots) => DepotStatus vs slots

Initial DepotStatus to start off the game.

initSomeDepotStatus :: (KnownNat vs, 1 <= vs) => SomeDepotStatus vs

Initial SomeDepotStatus to start off the game.

baseDepotCapacity :: forall vs slots. KnownNat vs => VehicleData vs -> DepotStatus vs slots -> Double

Total base capacity of all slots, in eggs per second.

totalDepotCapacity :: forall vs slots. KnownNat vs => VehicleData vs -> Bonuses -> DepotStatus vs slots -> Double

Total capacity of all vehicles, factoring in bonuses.

vehicleHistory :: KnownNat (slots + 1) => DepotStatus vs slots -> Map (Finite vs) (Finite (slots + 1))

How many of each vehicle has been purchased so far. If key is not found, zero purchases is implied.

vehiclePrice :: forall vs slots. (KnownNat vs, KnownNat slots) => VehicleData vs -> DepotStatus vs slots -> Finite vs -> Maybe Bock

Get the BASE price of a given vehicle, if a purchase were to be made. Does not check if purchase is legal (see upgradeVehicle).

upgradeVehicle :: (KnownNat vs, KnownNat slots) => VehicleData vs -> Bonuses -> Finite slots -> Finite vs -> DepotStatus vs slots -> Either (Finite vs) (Bock, DepotStatus vs slots)

Purchase a vehicle upgrade. Returns cost and new depot status, if purchase is valid.

Purchase is invalid if purchasing a vehicle in a slot where a greater vehicle is already purchased.

upgradeSomeVehicle :: forall vs. KnownNat vs => VehicleData vs -> Bonuses -> Integer -> Finite vs -> SomeDepotStatus vs -> Either (SomeVehicleUpgradeError vs) (Bock, SomeDepotStatus vs)

Purchase a vehicle upgrade. Returns cost and new depot status, if purchase is valid.

Purchase is invalid if purchasing a vehicle in a slot where a greater vehicle is already purchased, and also if there is no such slot.

vehicleUpgrades :: forall vs slots. KnownNat slots => VehicleData vs -> Bonuses -> DepotStatus vs slots -> ((Vector slots :.: Vector vs) :.: Either (Finite vs)) Bock

List all possible vehicle upgrades.

someVehicleUpgrades :: forall vs. VehicleData vs -> Bonuses -> SomeDepotStatus vs -> ((Map Integer :.: Vector vs) :.: Either (Finite vs)) Bock

List all possible vehicle upgrades for a SomeDepotStatus.